hw_05 Рефакторинг DAO и тестов

+++ hw5/src/main/java/ru/otus/boot/hw5/dao/author/AuthorDao.java
@@ -15,7 +15,7 @@ public interface AuthorDao {
 
     List<Author> getAll();
 
-    Author getById(Integer id);
+    Author getById(Long id);
 
     Author getByFirstAndLastName(String firstName, String lastname);
